Adjust autoCompleteTableView once auto complete terms get sorted
This is needed so that we can adjust the table view height to fit the auto complete list in case the count of terms is less than maximumNumberOfAutoCompleteRows. For example, if the maximumNumberOfAutoCompleteRows is 3, and the auto complete list contains just one term, we want to shrink the table view to just one row.